+++ AsteroidGame/VisualObjects/Asteroid.cs
@@ -7,20 +7,6 @@ using System.Threading.Tasks;
 
 namespace AsteroidGame.VisualObjects
 {
-    internal abstract class ImageObject : VisualObject
-    {
-        private readonly Image _Image;
-
-        protected ImageObject(Point Position, Point Direction, Size Size, Image Image) : base(Position, Direction, Size)
-        {
-            _Image = Image;
-        }
-
-        public override void Draw(Graphics g)
-        {
-            g.DrawImage(_Image, _Position.X, _Position.Y, _Size.Width, _Size.Height);
-        }
-    }
 
     internal class Asteroid : ImageObject
     {
