+++ AsteroidGame/VisualObjects/Bullet.cs
@@ -11,6 +11,7 @@ namespace AsteroidGame.VisualObjects
     {
         private const int __BulletSizeX = 20;
         private const int __BulletSizeY = 5;
+        private const int __BulletSpeed = 3;
 
         public Bullet(int Position) : base(new Point(0, Position), Point.Empty, new Size(__BulletSizeX, __BulletSizeY))
         {
@@ -27,7 +28,7 @@ namespace AsteroidGame.VisualObjects
 
         public override void Update()
         {
-            _Position = new Point(_Position.X + 3, _Position.Y);
+            _Position = new Point(_Position.X + __BulletSpeed, _Position.Y);
         }
     }
 }
